refactor: change RecordFileManipulator path
